local-console/pyproject.toml
[build-system]
requires = ["setuptools >= 64", "setuptools-scm >= 8"]
build-backend = "setuptools.build_meta"

[project]
name = "local-console"
description = "Local Console for offline camera operation"
license = {file = "LICENSE"}
dynamic = ["version"]
authors = [
	{name = "DevTools and SDKs team", email = "engineering@midokura.com"}
]
classifiers = [
    "License :: OSI Approved :: Apache Software License",
	"Natural Language :: English",
	"Framework :: Trio",
    "Operating System :: OS Independent",
	"Topic :: Scientific/Engineering :: Artificial Intelligence",
	"Topic :: Scientific/Engineering :: Image Recognition",
  	"Programming Language :: Python :: 3.10",
  	"Programming Language :: Python :: 3.11",
]
requires-python = ">=3.10"
dependencies = [
	"annotated-types==0.6.0",
	"async-generator==1.10",
	"asyncgui==0.6.1",
	"asynckivy==0.6.2",
	"attrs==23.2.0",
	"certifi==2024.2.2",
	"cffi==1.16.0",
	"charset-normalizer==3.3.2",
	"click==8.1.7",
	"cryptography==42.0.7",
	"decorator==5.1.1",
	"docutils==0.21.2",
	"exceptiongroup==1.2.1",
	"idna==3.7",
	"importlib_metadata==7.1.0",
	"Kivy==2.3.0",
	"Kivy-Garden==0.1.5",
	"kivymd @ git+https://github.com/kivymd/KivyMD.git@06e5f0c3330170200918e712345a2b4be6290c65",
	"materialyoucolor==2.0.9",
	"mypy-extensions==1.0.0",
	"outcome==1.3.0.post0",
	"packaging==24.0",
	"paho-mqtt==1.6.1",
	"pillow==10.3.0",
	"psutil==5.9.8",
	"py==1.11.0",
	"pycparser==2.22",
	"pycryptodome==3.20.0",
	"pydantic==2.7.1",
	"pydantic_core==2.18.2",
	"Pygments==2.17.2",
	"pypng==0.20220715.0",
	"qrcode==7.4.2",
	"requests==2.31.0",
	"retry==0.9.2",
	"sniffio==1.3.1",
	"sortedcontainers==2.4.0",
	"trio==0.25.0",
	"trio-typing==0.10.0",
	"trio-util==0.7.0",
	"typer==0.12.3",
	"types-docutils==0.21.0.20240423",
	"types-paho-mqtt==1.6.0.20240321",
	"types-Pygments==2.17.0.20240310",
	"types-retry==0.9.9.4",
	"types-setuptools==69.2.0.20240317",
	"typing_extensions==4.11.0",
	"urllib3==2.2.1",
	"zipp==3.18.1",
	"opencv-python==4.10.0.82",
	"numpy==1.26.4"
]

[project.urls]
# All options at https://packaging.python.org/en/latest/guides/writing-pyproject-toml/#urls
homepage = "https://github.com/midokura/local-console"

[tool.setuptools.dynamic]
version = {file = ["VERSION"]}

[tool.setuptools]
license-files = ["*LICENSE"]

[tool.setuptools.packages.find]
where = ["src"]

# Plugin architecture via entry points
[project.entry-points."local_console.plugin"]
base = "local_console.cli:PluginBase"
broker = "local_console.commands.broker:BrokerCommand"
config = "local_console.commands.config:ConfigCommand"
deploy = "local_console.commands.deploy:DeployCommand"
get = "local_console.commands.get:GetCommand"
gui = "local_console.commands.gui:GUICommand"
logs = "local_console.commands.logs:LogsCommand"
qr = "local_console.commands.qr:QRCommand"
rpc = "local_console.commands.rpc:RPCCommand"

[project.scripts]
local-console = "local_console.__main__:app"

# We will want to enable this once all errors can be
# displayed by the GUI itself, so we won't need the
# terminal to be spawned anymore. See:
# https://packaging.python.org/en/latest/guides/writing-pyproject-toml/#creating-executable-scripts
#[project.gui-scripts]
#local-console-gui = "local_console.__main__:app---to-be-defined"
